+++ tests/component/header.test.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import Header from "../../src/app/components/header";
+import Header from "../../src/app/components/header/header";
 import { describe, expect, it } from "@jest/globals";
 import { render } from "@testing-library/react";
